+++ src/Qrala.py
@@ -13,18 +13,17 @@
 #   Licence: GNU Affero General Public License v3.0                             #
 #################################################################################
 
-
-
+# Import Modules
+import VCard.VCardView as vcard
+import WIFI.WIFI_View as wifi
+import SystemDependency as sys_dep
+import ConstantStyle as cs
 
 # TKinter
+from tkinter import filedialog, Text, Tk, Menu
+from tkinter.constants import NORMAL, END, NONE, CENTER
+from tkinter.ttk import Frame, Label, Button, Combobox, Notebook
 
-from tkinter import *
-
-# remove following trey lines, because they are not used anymore
-import tkinter as tk
-from tkinter import filedialog
-from tkinter import ttk
-from tkinter import Entry
 # XML
 import xml.dom.minidom
 
@@ -37,18 +36,15 @@ from PIL import Image, ImageTk
 
 from pprint import pprint
 
-
-import SystemDependency as depend
-
-BG_COLOR = "#696969"
-secColor = "#b5b5b5"
-FONT_1 = ("Helvetica", 14)  # ("Century Gothic", 14, BOLD)
-FONT_2 = ("Helvetica", 16)
+BACKGROUND = cs.BACKGROUND
+SECONDARY = cs.SECONDARY
+FONT_1 = cs.FONT_1
+FONT_2 = cs.FONT_2
 
 
-OPERATING_SYSTEM = depend.getOS()
+OPERATING_SYSTEM = sys_dep.getOS()
 
-rootPath = depend.getRootPath()
+rootPath = sys_dep.getRootPath()
 print(rootPath)
 
 domtree = xml.dom.minidom.parse(
@@ -96,13 +92,7 @@ def get_From_XML(str):
     return tagname[0].firstChild.data
 
 
-def changeOnHover(button, colorOnHover, colorOnLeave):
 
-    button.bind("<Enter>", func=lambda e: button.config(
-        background=colorOnHover))
-
-    button.bind("<Leave>", func=lambda e: button.config(
-        background=colorOnLeave))
 
 def removePlaceholder(event, current_entry):
     current_entry.configure(state=NORMAL)
@@ -123,79 +113,18 @@ def get_QR():        # WIFI:S:password;T:WPA;P:ssid;; #formart
 
 
 
-# returns list for mac
-
-
-def get_WIFI_QR(note, new_bg_img):
-    wifi_QR = Frame(note)
-    wifi_QR.configure(background=BG_COLOR)
-
-    # GUI
-    label_In = tk.Label(wifi_QR, bg=BG_COLOR, font=FONT_2, fg="black",
-                        text=get_From_XML('QR_Inhalt'))
-    label_In.grid(column=0,
-                  row=2,
-                  padx=20,
-                  pady=8)
-
-    # Check for OS
-    if OPERATING_SYSTEM == "darwin":
-        # MAC OS
-        print("\nOS:\t", "Mac OS\n")
-
-    elif OPERATING_SYSTEM == "win64":
-        # Windows 64-bit
-        print("\nOS:\t", "Windows\n")
-
-    # Label
-    label_In = tk.Label(wifi_QR, bg=BG_COLOR, font=FONT_2, fg="black",
-                        text='Enter your WIFI information')
-    label_In.grid(column=0,
-                  row=0,
-                  padx=20,
-                  pady=8)
-
-    wifi_name = Entry(wifi_QR, width=30)
-    wifi_name.grid(column=0,
-                        row=1,
-                        padx=20,
-                        pady=8)
-    wifi_name.insert(0, 'SSID (often WIFI-Name)')
-    wifi_name.configure(state=DISABLED)
-    wifi_name.bind("<Button-1>", lambda event: removePlaceholder(event, wifi_name))
-
-    wifi_password = Entry(wifi_QR, width=30)
-    wifi_password.grid(column=0,
-                        row=2,
-                        padx=20,
-                        pady=8)
-    wifi_password.insert(0, 'Password')
-    wifi_password.configure(state=DISABLED)
-    wifi_password.bind("<Button-1>", lambda event: removePlaceholder(event, wifi_password))
-
-    # Button zum QR-Code Generieren
-    get_QR_button = tk.Button(wifi_QR, text=get_From_XML('Generate_Code'), highlightbackground=BG_COLOR, padx=4,
-                              pady=2, font=FONT_1, command=get_QR)
-    get_QR_button.grid(column=0,
-                       row=4,
-                       padx=10,
-                       pady=18)
-    changeOnHover(get_QR_button, "white", secColor)
-
-    # Background Img
-    img_label = Label(wifi_QR, image=new_bg_img, bg=BG_COLOR)
-    img_label.place(x=680, y=306)
-
-    return wifi_QR
 
 
 def getCustomQR(note, new_bg_img):
     customQR = Frame(note)
-    customQR.configure(background=BG_COLOR)
+    # customQR.configure(background=BACKGROUND)
 
     # Label
-    label_In = tk.Label(customQR, bg=BG_COLOR, font=FONT_2, fg="black",
-                        text=get_From_XML('QR_Inhalt'))
+    label_In = Label(customQR,
+                     background=BACKGROUND,
+                     font=FONT_2,
+                     foreground="black",
+                     text=get_From_XML('QR_Inhalt'))
     label_In.grid(column=0,
                   row=2,
                   padx=20,
@@ -203,163 +132,54 @@ def getCustomQR(note, new_bg_img):
 
     # Textbox
     global inputText
-    inputText = Text(customQR, height=20, width=60, bg=secColor, font=FONT_2)
+    inputText = Text(customQR, height=20, width=60, bg=SECONDARY, font=FONT_2)
     inputText.grid(column=0,
                    row=3,
                    padx=20, sticky="nesw")
 
     # Button zum QR-Code Generieren
-    getQrButton = tk.Button(customQR, text=get_From_XML('Generate_Code'), highlightbackground=BG_COLOR, padx=4,
-                            pady=2, font=FONT_1, command=get_QR)
-    changeOnHover(getQrButton, "white", secColor)
+    getQrButton = Button(customQR,
+                         text=get_From_XML('Generate_Code'),
+                         # highlightbackground=BG_COLOR,
+                         # padx=4,
+                         # pady=2,
+                         # font=FONT_1,
+                         command=get_QR)
+    cs.changeOnHover(getQrButton, "white", SECONDARY)
     getQrButton.grid(column=0,
                      row=4,
                      padx=10,
                      pady=18)
 
     # Background Img
-    img_label = Label(customQR, image=new_bg_img, bg=BG_COLOR)
+    img_label = Label(customQR, image=new_bg_img, background=BACKGROUND)
     img_label.place(x=680, y=306)
 
     return customQR
 
 
 
-def getContactQR(note, new_bg_img):
-    contact_QR = Frame(note)
-    contact_QR.configure(background=BG_COLOR)
 
-    # Background Img
-    img_label = Label(contact_QR, image=new_bg_img, bg=BG_COLOR)
-    img_label.place(x=680, y=306)
-
-    # Label
-    label_In = tk.Label(contact_QR, bg=BG_COLOR, font=FONT_2, fg="black",
-                        text='Enter your contact information for the VCard')
-    label_In.grid(column=0,
-                  row=1,
-                  padx=20,
-                  pady=8)
-
-    showed_name = Entry(contact_QR, width=30)
-    showed_name.grid(column=0,
-                        row=2,
-                        padx=20,
-                        pady=8)
-    showed_name.insert(0, 'Showed Name')
-    showed_name.configure(state=DISABLED)
-    showed_name.bind("<Button-1>", lambda event: removePlaceholder(event, showed_name))
-
-    last_name = Entry(contact_QR, width=30)
-    last_name.grid(column=0,
-                        row=3,
-                        padx=20,
-                        pady=8)
-    last_name.insert(0, 'Last Name')
-    last_name.configure(state=DISABLED)
-    last_name.bind("<Button-1>", lambda event: removePlaceholder(event, last_name))
-
-    first_name = Entry(contact_QR, width=30)
-    first_name.grid(column=1,
-                        row=3,
-                        padx=20,
-                        pady=8)
-    first_name.insert(0, 'First Name')
-    first_name.configure(state=DISABLED)
-    first_name.bind("<Button-1>", lambda event: removePlaceholder(event, first_name))
-
-    # Tel Number Entry
-    tel_number = Entry(contact_QR, width=30)
-    tel_number.grid(column=0,
-                        row=4,
-                        padx=20,
-                        pady=8)
-    tel_number.insert(0, 'Tel. Number')
-    tel_number.configure(state=DISABLED)
-    tel_number.bind("<Button-1>", lambda event: removePlaceholder(event, tel_number))
-
-    # Email Entry
-    email = Entry(contact_QR, width=30)
-    email.grid(column=1,
-                        row=4,
-                        padx=20,
-                        pady=8)
-    email.insert(0, 'Email')
-    email.configure(state=DISABLED)
-    email.bind("<Button-1>", lambda event: removePlaceholder(event, email))
-
-    # organization Entry
-    organization = Entry(contact_QR, width=30)
-    organization.grid(column=0,
-                        row=5,
-                        padx=20,
-                        pady=8)
-    organization.insert(0, 'Organization')
-    organization.configure(state=DISABLED)
-    organization.bind("<Button-1>", lambda event: removePlaceholder(event, organization))
-
-    # address Entry
-    address = Entry(contact_QR, width=30)
-    address.grid(column=1,
-                        row=5,
-                        padx=20,
-                        pady=8)
-    address.insert(0, 'Address')
-    address.configure(state=DISABLED)
-    address.bind("<Button-1>", lambda event: removePlaceholder(event, address))
-
-    # city Entry
-    city = Entry(contact_QR, width=30)
-    city.grid(column=0,
-                        row=6,
-                        padx=20,
-                        pady=8)
-    city.insert(0, 'City')
-    city.configure(state=DISABLED)
-    city.bind("<Button-1>", lambda event: removePlaceholder(event, city))
-
-    # country Entry
-    country = Entry(contact_QR, width=30)
-    country.grid(column=1,
-                        row=6,
-                        padx=20,
-                        pady=8)
-    country.insert(0, 'Country')
-    country.configure(state=DISABLED)
-    country.bind("<Button-1>", lambda event: removePlaceholder(event, country))
-
-    # postalcode Entry
-    postalcode = Entry(contact_QR, width=30)
-    postalcode.grid(column=0,
-                        row=7,
-                        padx=20,
-                        pady=8)
-    postalcode.insert(0, 'Postalcode')
-    postalcode.configure(state=DISABLED)
-    postalcode.bind("<Button-1>", lambda event: removePlaceholder(event, postalcode))
-
-
-    return contact_QR
 
 
 
 def get_Settings(note, new_bg_img):
     settings = Frame(note)
-    settings.configure(background=BG_COLOR)
+    # settings.configure(background=BG_COLOR)
 
     # Background Img
-    img_label = Label(settings, image=new_bg_img, bg=BG_COLOR)
+    img_label = Label(settings, image=new_bg_img, background=BACKGROUND)
     img_label.place(x=680, y=306)
 
-    langLabel = ttk.Label(settings, text="Language:")
-    langLabel.configure(background=BG_COLOR, font=FONT_1, padding=10)
+    langLabel = Label(settings, text="Language:")
+    langLabel.configure(background=BACKGROUND, font=FONT_1, padding=10)
     langLabel.grid(column=0, row=2)
 
-    chooseLang = ttk.Combobox(settings,
+    chooseLang = Combobox(settings,
                               values=["Deutsch",
                                       "English"])
 
-    chooseLang.configure(background=BG_COLOR, font=FONT_1,
+    chooseLang.configure(background=BACKGROUND, font=FONT_1,
                          takefocus=NONE, justify=CENTER)
     chooseLang.grid(column=1, row=2)
     chooseLang.current(1)
@@ -380,7 +200,7 @@ def main():
 
     # Menubar
     menubar = Menu(win)
-    win.configure(background=BG_COLOR, menu=menubar)
+    win.configure(background=BACKGROUND, menu=menubar)
 
     # File Menu
     filemenu = Menu(menubar)
@@ -400,16 +220,16 @@ def main():
     setting_Icon = ImageTk.PhotoImage(setting_16px)
 
     # Notebook
-    note = ttk.Notebook(win)
+    note = Notebook(win)
     note.pack(fill="both", expand=1)
 
     customQR = getCustomQR(note, new_bg_img)
     note.add(customQR, text="Custom")
 
-    wifiQR = get_WIFI_QR(note, new_bg_img)
+    wifiQR = wifi.getFrame(note, new_bg_img)
     note.add(wifiQR, text="WIFI", image=wifi_Icon, compound="left")
 
-    contactQR = getContactQR(note, new_bg_img)
+    contactQR = vcard.getFrame(note, new_bg_img)
     note.add(contactQR, text="Contact", image=contact_Icon, compound="left")
 
     settings = get_Settings(note, new_bg_img)