+++ changelog.md
@@ -1,11 +1,19 @@
 <h1> Changelog </h1>
 
 > Qrala &bull; _since 2021_ &bull;  <br/> 
-> Version: 0.9.4 <br/> 
+> Version: 0.9.5 <br/> 
 > All rights reserved. <br/>
 > Copyright &copy; 2021 - 2022. [Christoph Rohde](https://github.com/CodebyCR) <br/>
 > Licence: GNU Affero General Public License v3.0
 #
+
+---
+<h3>Version 0.9.5)  </h3>
+- <b>[WIFI QR-Codes are now working on macOS.](/issues/3)</b>
+- <b>[add better file structure.](/issues/1)</b>
+- add some code for VCard QR-Codes
+- add some code for WIFI QR-Codes
+- some readability improvements and cleanups
 
 ---
 <h3>Version 0.9.4)  </h3>
@@ -13,8 +21,8 @@
 - add .gitignore
 - add some code for VCard QR-Codes
 - add some code for WIFI QR-Codes
+- 
 - some readability improvements
-
 
 ---
 <h3>Version 0.9.2)  </h3>